Issue #739: move kernel to a separate file
