+++ deepreg/loss/label.py
@@ -5,9 +5,9 @@ from typing import List, Optional
 import tensorflow as tf
 
 from deepreg.constant import EPS
-from deepreg.loss.util import NegativeLossMixin, cauchy_kernel1d
-from deepreg.loss.util import gaussian_kernel1d_sigma as gaussian_kernel1d
-from deepreg.loss.util import separable_filter
+from deepreg.loss.kernel import cauchy_kernel1d
+from deepreg.loss.kernel import gaussian_kernel1d_sigma as gaussian_kernel1d
+from deepreg.loss.util import NegativeLossMixin, separable_filter
 from deepreg.registry import REGISTRY
